+++ lib/schleuder/cli.rb
@@ -63,14 +63,14 @@ module Schleuder
       end
     end
 
-    desc 'refresh_keys', "Refresh all keys of all list from the keyservers sequentially (one by one). (This is supposed to be run from cron weekly.)"
-    def refresh_keys
-      work_on_lists(:refresh_keys)
+    desc 'refresh_keys [list1@example.com]', "Refresh all keys of all list from the keyservers sequentially (one by one or on the passed list). (This is supposed to be run from cron weekly.)"
+    def refresh_keys(list=nil)
+      work_on_lists(:refresh_keys,list)
     end
 
-    desc 'pin_keys', "Find keys for subscriptions without a pinned key and try to pin a certain key"
-    def pin_keys
-      work_on_lists(:pin_keys)
+    desc 'pin_keys [list1@example.com]', "Find keys for subscriptions without a pinned key and try to pin a certain key (one by one or based on the passed list)."
+    def pin_keys(list=nil)
+      work_on_lists(:pin_keys,list)
     end
 
     desc 'install', "Set-up or update Schleuder environment (create folders, copy files, fill the database)."
@@ -303,8 +303,13 @@ Please notify the users and admins of this list of these changes.
       end
     end
     private
-    def work_on_lists(subj)
-      List.all.each do |list|
+    def work_on_lists(subj, list=nil)
+      selected_lists = if list.nil?
+        List.all
+      else
+        List.where(email: list)
+      end
+      selected_lists.each do |list|
         I18n.locale = list.language
         output = list.send(subj)
         if output.present?